+++ kernel/kernel/fs/vfs.c
@@ -59,7 +59,7 @@ void vfs_install()
 }
 devman_register_driver_installation(vfs_install);
 
-static void vfs_recieve_notification(uint32_t msg, uint32_t param)
+static void vfs_recieve_notification(uintptr_t msg, uintptr_t param)
 {
     switch (msg) {
     case DEVMAN_NOTIFICATION_NEW_DEVICE:
